+++ src/props.ts
@@ -1,4 +1,4 @@
-import { FieldPropsEnum } from "./models/FieldProps";
+import { FieldPropsEnum, FieldPropsOccurrence, SeparatedPropsMode } from "./models/FieldProps";
 
 export interface PropsGroupsInterface {
   editable: string[];
@@ -8,7 +8,7 @@ export interface PropsGroupsInterface {
   validation: string[];
   exceptions: string[];
   separated: string[];
-  types: {
+  occurrences: {
     [index: string]: "some" | "every";
   };
 }
@@ -62,22 +62,29 @@ export const props: PropsGroupsInterface = {
     FieldPropsEnum.disabled,
   ],
   separated: [
-    "values",
-    "initials",
-    "defaults",
-    "labels",
-    "placeholders",
-    "disabled",
-    "deleted",
-    "related",
-    "options",
-    "extra",
-    "bindings",
-    "types",
-    "hooks",
-    "handlers",
-    "autoFocus",
-    "refs"
+    SeparatedPropsMode.values,
+    SeparatedPropsMode.labels,
+    SeparatedPropsMode.placeholders,
+    SeparatedPropsMode.defaults,
+    SeparatedPropsMode.initials,
+    SeparatedPropsMode.disabled,
+    SeparatedPropsMode.deleted,
+    SeparatedPropsMode.types,
+    SeparatedPropsMode.related,
+    SeparatedPropsMode.rules,
+    SeparatedPropsMode.options,
+    SeparatedPropsMode.bindings,
+    SeparatedPropsMode.extra,
+    SeparatedPropsMode.hooks,
+    SeparatedPropsMode.handlers,
+    SeparatedPropsMode.validatedWith,
+    SeparatedPropsMode.validators,
+    SeparatedPropsMode.observers,
+    SeparatedPropsMode.interceptors,
+    SeparatedPropsMode.input,
+    SeparatedPropsMode.output,
+    SeparatedPropsMode.autoFocus,
+    SeparatedPropsMode.refs,
   ],
   functions: [
     FieldPropsEnum.observers,
@@ -94,19 +101,19 @@ export const props: PropsGroupsInterface = {
     FieldPropsEnum.isDirty,
     FieldPropsEnum.isPristine
   ],
-  types: {
-    isDirty: "some",
-    isPristine: "every",
-    isDefault: "every",
-    isValid: "every",
-    isEmpty: "every",
-    hasError: "some",
-    focused: "some",
-    blurred: "some",
-    touched: "some",
-    deleted: "every",
-    disabled: "every",
-    clearing: "every",
-    resetting: "every",
+  occurrences: {
+    isDirty: FieldPropsOccurrence.some,
+    isPristine: FieldPropsOccurrence.every,
+    isDefault: FieldPropsOccurrence.every,
+    isValid: FieldPropsOccurrence.every,
+    isEmpty: FieldPropsOccurrence.every,
+    hasError: FieldPropsOccurrence.some,
+    focused: FieldPropsOccurrence.some,
+    blurred: FieldPropsOccurrence.some,
+    touched: FieldPropsOccurrence.some,
+    deleted: FieldPropsOccurrence.every,
+    disabled: FieldPropsOccurrence.every,
+    clearing: FieldPropsOccurrence.every,
+    resetting: FieldPropsOccurrence.every,
   },
 };